requirements.txt
django==1.10.5
pyopenssl
requests
chatbotAI
lxml
wikipedia